requirements.txt
# Core web framework and real-time communication
Flask==3.0.0
Flask-SocketIO==5.3.6
python-socketio==5.10.0

# AI and agent framework
google-generativeai==0.3.2
strands-agents==0.1.0

# Model Context Protocol
mcp==1.18.0

# Environment and configuration
python-dotenv==1.0.0

# HTTP client for API calls
requests==2.31.0

# Async support
asyncio==3.4.3

# JSON handling
jsonschema==4.20.0

# Logging and utilities
colorlog==6.8.0

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1